Rehydrate vectors at the deepest available object in the JSON path
This change ensures that when restoring vectors removed from the source, we locate the deepest existing object along the JSON path to rehydrate them accurately.
